BOL-375 refactor: LoginType package 이동
